Change GinConfigSaverHook to use tensorboard.summary.text_pb()
This prevents breakage in an upcoming CL that moves the `text_summary` module and hides the PLUGIN_NAME constant.  As a replacement, we use the `text_pb()` method from TensorBoard's own summary API that is the equivalent of the text summary op but generates the Summary proto directly without any op execution.

The tag needs to be manually overridden because text_pb() currently adds a "/text_summary" suffix, but we're planning on removing that in the near future, so I have a TODO to myself to remove the override then.

PiperOrigin-RevId: 219666504