[docs] Include example of `gin.parse_config_file`
As a first-time user of gin, it seemed like gin might automatically detect and parse my `config.gin` file.  But it appears that we're actually required to call `gin.parse_config_file`, so I think it would help to explicitly include an example of that in the documentation of basic usage.