Add proper support for gin.REQUIRED in function signatures.
Previously, it had been a design decision to only support gin.REQUIRED at the call site. The primary reason for this was to minimize the extent to which functions were dependent on Gin. However, using gin.REQUIRED as a default value didn't throw an error (since it had previously been implemented simply as a string), and it seems many people having been using it in signatures, despite a lack of proper support. This CL adds proper support for this use case.

Also:
- Does a bit of cleanup around caching
- Handles builtin functions properly
PiperOrigin-RevId: 239895086