Fixes thread safety issues in scope management.
Currently, for configurable functions that are called within a scope, the scope is usually applied via a `scoping_wrapper` that:

  1. enters the appropriate scope (mutating the global _ACTIVE_SCOPES);
  2. calls the function;
  3. exits the scope.

In a multithreaded scenario, if _ACTIVE_SCOPES gets modified by another thread after step 1 but before step 2 (just by the other thread calling a scoped function that follows the same steps above), the function may be called in the wrong scope, with serious consequences (wrong config values applied).

To fix this, this CL makes scope management thread local. It also addresses one other potential thread safety concern by adding a lock around updates to the operative config. Other accesses of global variables in the gin wrapper are read only and should be thread safe (assuming the config itself isn't being mutated in another thread, but that would require unlocking the config and willful evilness by the user).

For testing, I wrote a test which exercise the scope management in a multithreaded scenario and reliably fails without the thread-local scope management this CL introduces.

PiperOrigin-RevId: 259978007