Make Gin dynamic class wrapping compatible with pickle.
A new take on how to do this, since the wrapt approach also didn't end up working. This approach does the following:

When wrapping a class where mutation should be disallowed, this dynamically subclasses the class's metaclass, and intercepts/wraps its __call__.
However, this metaclass then returns an instance of the original type, which allows pickling to work.
To preserve proper signature checking, there are a few changes also to the general Gin wrapper function, to always use the signature associated with the right __init__/__new__ function.

PiperOrigin-RevId: 368536947