Add class method support to dynamic registration.
In dynamic registration, configuring a class or method registers it, however the current method support requires methods to be registered before the class (this is to avoid overriding class methods with configurable versions unless necessary, since doing this breaks pickle compatibility). This is natural when using an `@gin.register` decorator, but for dynamic registration its important to allow the registration to occur in any order.

To address this, this change enables re-registration of functions/classes (as long as the same underlying function is being re-registered), and if a method is registered its parent class is re-registered. Any configurable references in the existing _CONFIG are updated to point to the newly re-registered version of the class.

PiperOrigin-RevId: 374299096