Add support for dynamic registration when generating config strings.
The basic strategy is to make the outputted configurations use dynamic registration if it was enabled in any parsed config. To preserve naming, an `import_source` field is added to the Configurable class, and an ImportManager class is used to prevent naming collisions when outputting import statements in cases where dynamic registration is enabled.

PiperOrigin-RevId: 404611529