Fix Gin parser test in Python 3.12.
It looks like in Python 3.12, certain exceptions that used to be SyntaxError instances are now tokenize.TokenError instances instead. The config parser test was validating that broken syntax was appropriately throwing errors during parsing... the specific exception type being thrown isn't particularly relevant, as long as it surfaces clearly to the user. So the fix just expands the list of acceptable error types allowed by the test.

PiperOrigin-RevId: 741539841